+++ src/utils/helpers.ts
@@ -1,4 +1,6 @@
-import { IntegrationResult } from "./types";
+import path = require("path");
+import fs = require("fs");
+import { IntegrationResult } from "../types/apiTypes";
 
 export function debounce<Params extends unknown[]>(func: (...args: Params) => unknown, timeout: number): (...args: Params) => void {
     let timer: NodeJS.Timeout;
@@ -18,4 +20,14 @@ export function unescapeResponseBody(result: IntegrationResult): string {
     // NOTE: body is a json escaped string (eg { \"node\": 123 })
     const unescapedBody = result.body.replace(/\\"/g, "\"");
     return unescapedBody;
+}
+
+export function findDirectoryWithinParent(currentFolder: string, folderNameToLookFor: string, maxUp: number) : string | undefined  {
+    const pathCheck = path.join(currentFolder, folderNameToLookFor);
+    if (maxUp === 0) { return undefined; }
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-call
+    if (fs.existsSync(pathCheck)) { return pathCheck; }
+    else {
+        return findDirectoryWithinParent(path.dirname(currentFolder), folderNameToLookFor, --maxUp);
+    }
 }